🎨 (app) Structured Models Code
