+++ libs/app/ingredient/utils/src/interfaces/ingredient.interface.ts
@@ -1,4 +1,10 @@
 export interface IIngredient {
-    ingredientId: string;
+    id: string;
     name: string;
-}
+    tags?: string[];
+  }
+  
+  export interface QuantityIngredient extends IIngredient {
+    quantity: number;
+    scale: string;
+  }
