+++ libs/app/recipe/feature/src/recipe.page.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { recipeList } from "@fridge-to-plate/app/recipe/data-access";
-import { IRecipe} from "../../../../../models/irecipe";
+import { IRecipe} from "@fridge-to-plate/app/recipe/utils";
 
 @Component({
   selector: 'recipe-page',
