+++ libs/app/recipe/utils/src/index.ts
@@ -1,1 +1,1 @@
-export * from './recipe.module';
+export * from './interfaces';
