+++ libs/app/recipe/utils/src/interfaces/index.ts
@@ -1,0 +1,2 @@
+export * from './recipe-step.interface';
+export * from './recipe.interface';
