+++ libs/app/recipe/utils/src/interfaces/recipe-step.interface.ts
@@ -1,0 +1,5 @@
+export interface IRecipeStep {
+    instructionHeading: string;
+    instructionBody: string;
+    stepDuration?: number;
+}
