+++ libs/app/recipe/utils/src/interfaces/recipe.interface.ts
@@ -1,8 +1,17 @@
-import { IIngredient } from "@fridge-to-plate/app/ingredient/utils";
+import { IIngredient } from '@fridge-to-plate/app/ingredient/utils';
+import { IRecipeStep } from './recipe-step.interface';
 
 export interface IRecipe {
-    recipeId: string;
-    name: string;
-    ingredients: IIngredient[];
-    instructions: String[];
+  id: string;
+  name: string;
+  recipeImage: string;
+  ingredients: IIngredient[];
+  steps: IRecipeStep[];
+  rating?: number;
+  difficulty: 'easy' | 'medium' | 'hard';
+  meta?: {
+    prepTime?: number;
+    numberOfServings?: number;
+    tags: string[];
+  };
 }
