--- models/ingredient.ts
@@ -1,10 +1,0 @@
-export interface Ingredient {
-  id: string;
-  name: string;
-  tags?: string[];
-}
-
-export interface UserIngredient extends Ingredient {
-  quantity?: number;
-  mass?: number;
-}
