--- models/irecipe-step.ts
@@ -1,5 +1,0 @@
-export interface IRecipeStep {
-  instructionHeading: string;
-  instructionBody: string;
-  stepDuration?: number;
-}
